refactor: separate Vulkan RHI into files
